+++ src/senaite/core/setuphandlers.py
@@ -187,7 +187,6 @@ def install(context):
     add_senaite_setup(portal)
     add_senaite_setup_items(portal)
     add_dexterity_portal_items(portal)
-    add_dexterity_setup_items(portal)
 
     # Set CMF Form actions
     setup_form_controller_actions(portal)
@@ -197,30 +196,6 @@ def install(context):
     setup_markup_schema(portal)
 
     logger.info("SENAITE CORE install handler [DONE]")
-
-
-def add_dexterity_setup_items(portal):
-    """Adds the Dexterity Container in the Setup Folder
-
-    N.B.: We do this in code, because adding this as Generic Setup Profile in
-          `profiles/default/structure` flushes the contents on every import.
-    """
-    # Tuples of ID, Title, FTI
-    items = [
-        ("dynamic_analysisspecs",  # ID
-         "Dynamic Analysis Specifications",  # Title
-         "DynamicAnalysisSpecs"),  # FTI
-
-        ("interpretation_templates",
-         "Interpretation Templates",
-         "InterpretationTemplates"),
-
-        ("sample_containers",
-         "Sample Containers",
-         "SampleContainers")
-    ]
-    setup = api.get_setup()
-    add_dexterity_items(setup, items)
 
 
 def add_senaite_setup(portal):
@@ -260,6 +235,13 @@ def add_senaite_setup_items(portal):
         ("samplingdeviations", "Sampling Deviations", "SamplingDeviations"),
         ("batchlabels", "Batch Labels", "BatchLabels"),
         ("analysiscategories", "Analysis Categories", "AnalysisCategories"),
+        ("instrumentlocations", "Instrument Locations", "InstrumentLocations"),
+        ("samplecontainers", "Sample Containers", "SampleContainers"),
+        ("attachmenttypes", "Attachment Types", "AttachmentTypes"),
+        ("dynamicanalysisspecs", "Dynamic Analysis Specifications",
+         "DynamicAnalysisSpecs"),
+        ("interpretationtemplates", "Interpretation Templates",
+         "InterpretationTemplates"),
         ("suppliers", "Suppliers", "Suppliers"),
     ]
     setup = api.get_senaite_setup()