+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2573 Move Interpretation Templates to Senaite setup folder
+- #2563 Move Dynamic Analysis Specs to Setup folder
+- #2554 Migrate AttachmentTypes to Dexterity
+- #2572 Move Sample Containers to Senaite setup folder
+- #2571 Move InstrumentLocations to Senaite setup folder
+- #2578 Fix setup DX folders are not indexed in `uid_catalog`
 - #2576 Fix analyses and services are re-indexed after modifying category
 - #2577 Fix content types typo
 - #2567 Migrate AnalysisCategories to Dexterity